+++ src/data_plane/states/segment/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod cache;
+pub(crate) mod record;
 pub(crate) mod tracker;
 use cache::*;
