+++ src/data_plane/wal.rs
@@ -1,13 +1,117 @@
 use std::collections::VecDeque;
 use std::fs::{self, File, OpenOptions};
-use std::io::{self, BufWriter, Write};
+use std::io::{self, BufWriter, Read, Write};
 use std::path::PathBuf;
 
+use bytes::Bytes;
+
 const DEFAULT_MAX_FILE_SIZE: u64 = 64 * 1024 * 1024; // 64MB
+const HEADER_SIZE: usize = 9; // crc32(4) + type(1) + length(4)
+const TRAILING_LENGTH_SIZE: usize = 4;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[repr(u8)]
+pub(crate) enum WalRecordType {
+    Data = 0,
+    BatchEnd = 1,
+}
+
+impl WalRecordType {
+    fn from_u8(v: u8) -> io::Result<Self> {
+        match v {
+            0 => Ok(WalRecordType::Data),
+            1 => Ok(WalRecordType::BatchEnd),
+            _ => Err(io::Error::new(
+                io::ErrorKind::InvalidData,
+                "unknown record type",
+            )),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct WalRecord {
+    pub record_type: WalRecordType,
+    pub payload: Bytes,
+}
+
+impl WalRecord {
+    pub(crate) fn data(payload: Bytes) -> Self {
+        WalRecord {
+            record_type: WalRecordType::Data,
+            payload,
+        }
+    }
+
+    pub(crate) fn batch_end() -> Self {
+        WalRecord {
+            record_type: WalRecordType::BatchEnd,
+            payload: Bytes::new(),
+        }
+    }
+
+    pub(crate) fn encoded_size(&self) -> usize {
+        HEADER_SIZE + self.payload.len() + TRAILING_LENGTH_SIZE
+    }
+
+    pub(crate) fn encode_to(&self, writer: &mut impl Write) -> io::Result<()> {
+        let payload_len = self.payload.len() as u32;
+
+        let mut hasher = crc32fast::Hasher::new();
+        hasher.update(&[self.record_type as u8]);
+        hasher.update(&payload_len.to_be_bytes());
+        hasher.update(&self.payload);
+        let crc = hasher.finalize();
+
+        writer.write_all(&crc.to_be_bytes())?;
+        writer.write_all(&[self.record_type as u8])?;
+        writer.write_all(&payload_len.to_be_bytes())?;
+        writer.write_all(&self.payload)?;
+        writer.write_all(&payload_len.to_be_bytes())?;
+        Ok(())
+    }
+
+    pub(crate) fn decode_from(reader: &mut impl Read) -> io::Result<Self> {
+        let mut header = [0u8; HEADER_SIZE];
+        reader.read_exact(&mut header)?;
+
+        let stored_crc = u32::from_be_bytes(header[0..4].try_into().unwrap());
+        let record_type = WalRecordType::from_u8(header[4])?;
+        let payload_len = u32::from_be_bytes(header[5..9].try_into().unwrap()) as usize;
+
+        let mut payload = vec![0u8; payload_len];
+        reader.read_exact(&mut payload)?;
+
+        let mut trailing_len = [0u8; TRAILING_LENGTH_SIZE];
+        reader.read_exact(&mut trailing_len)?;
+        let trailing = u32::from_be_bytes(trailing_len);
+        if trailing as usize != payload_len {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidData,
+                "trailing length mismatch",
+            ));
+        }
+
+        let mut hasher = crc32fast::Hasher::new();
+        hasher.update(&[record_type as u8]);
+        hasher.update(&(payload_len as u32).to_be_bytes());
+        hasher.update(&payload);
+        let computed_crc = hasher.finalize();
+
+        if stored_crc != computed_crc {
+            return Err(io::Error::new(io::ErrorKind::InvalidData, "CRC mismatch"));
+        }
+
+        Ok(WalRecord {
+            record_type,
+            payload: Bytes::from(payload),
+        })
+    }
+}
 
 pub trait WalStorage {
-    fn write_batch(&mut self, records: &[u8]) -> io::Result<u64>;
-    fn fsync(&mut self) -> io::Result<()>;
+    fn buf(&mut self) -> &mut Vec<u8>;
+    fn flush_batch(&mut self) -> io::Result<u64>;
     fn maybe_rotate(&mut self) -> io::Result<()>;
     fn delete_below(&mut self, watermark_lsn: u64);
 
@@ -22,6 +126,7 @@ pub struct WalWriter {
     data_dir: PathBuf,
     next_seq: u64,
     active_writer: Option<BufWriter<File>>,
+    batch_buf: Vec<u8>,
 }
 
 struct WalFileEntry {
@@ -44,6 +149,7 @@ impl WalWriter {
             data_dir,
             next_seq: 1,
             active_writer: None,
+            batch_buf: Vec::new(),
         };
         writer.open_new_file()?;
         Ok(writer)
@@ -110,6 +216,7 @@ impl WalWriter {
             data_dir,
             next_seq: entries.last().map(|(max_seq, _)| *max_seq).unwrap_or(0) + 1,
             active_writer: None,
+            batch_buf: Vec::new(),
         };
 
         if writer.files.is_empty() {
@@ -137,33 +244,31 @@ impl WalWriter {
 }
 
 impl WalStorage for WalWriter {
-    fn write_batch(&mut self, records: &[u8]) -> io::Result<u64> {
-        let lsn = self.next_lsn;
-
-        let writer = self
-            .active_writer
-            .as_mut()
-            .expect("invariant: active writer exists");
-        writer.write_all(records)?;
-
-        let entry = self.active_entry_mut();
-        entry.size += records.len() as u64;
-        entry.max_lsn = lsn;
-
-        self.next_lsn += 1;
-
-        Ok(lsn)
+    fn buf(&mut self) -> &mut Vec<u8> {
+        &mut self.batch_buf
     }
 
-    fn fsync(&mut self) -> io::Result<()> {
+    fn flush_batch(&mut self) -> io::Result<u64> {
+        let lsn = self.next_lsn;
+        let batch_len = self.batch_buf.len() as u64;
+
         let writer = self
             .active_writer
             .as_mut()
             .expect("invariant: active writer exists");
+        writer.write_all(&self.batch_buf)?;
         writer.flush()?;
         writer.get_ref().sync_all()?;
 
-        self.maybe_rotate()
+        let entry = self.active_entry_mut();
+        entry.size += batch_len;
+        entry.max_lsn = lsn;
+
+        self.batch_buf.clear();
+        self.next_lsn += 1;
+
+        self.maybe_rotate()?;
+        Ok(lsn)
     }
 
     fn maybe_rotate(&mut self) -> io::Result<()> {
@@ -245,6 +350,7 @@ pub mod tests {
         pub rotations: u32,
         pub deletes: Vec<u64>,
         next_lsn: u64,
+        batch_buf: Vec<u8>,
     }
 
     impl MockWalStorage {
@@ -255,21 +361,23 @@ pub mod tests {
                 rotations: 0,
                 deletes: Vec::new(),
                 next_lsn: 1,
+                batch_buf: Vec::new(),
             }
         }
     }
 
     impl WalStorage for MockWalStorage {
-        fn write_batch(&mut self, records: &[u8]) -> io::Result<u64> {
-            let lsn = self.next_lsn;
-            self.writes.push((records.to_vec(), lsn));
-            self.next_lsn += 1;
-            Ok(lsn)
+        fn buf(&mut self) -> &mut Vec<u8> {
+            &mut self.batch_buf
         }
 
-        fn fsync(&mut self) -> io::Result<()> {
+        fn flush_batch(&mut self) -> io::Result<u64> {
+            let lsn = self.next_lsn;
+            self.writes.push((self.batch_buf.clone(), lsn));
+            self.batch_buf.clear();
             self.fsyncs += 1;
-            Ok(())
+            self.next_lsn += 1;
+            Ok(lsn)
         }
 
         fn maybe_rotate(&mut self) -> io::Result<()> {
@@ -284,18 +392,21 @@ pub mod tests {
         fn assert_invariants(&self) {}
     }
 
+    fn wal_write(wal: &mut WalWriter, data: &[u8]) -> io::Result<u64> {
+        wal.buf().extend_from_slice(data);
+        wal.flush_batch()
+    }
+
     #[test]
     fn wal_writer_basic_lifecycle() {
         let dir = tempfile::tempdir().unwrap();
         let mut wal = WalWriter::new(dir.path().to_path_buf()).unwrap();
 
-        let lsn1 = wal.write_batch(b"record1").unwrap();
+        let lsn1 = wal_write(&mut wal, b"record1").unwrap();
         assert_eq!(lsn1, 1);
-        wal.fsync().unwrap();
 
-        let lsn2 = wal.write_batch(b"record2").unwrap();
+        let lsn2 = wal_write(&mut wal, b"record2").unwrap();
         assert_eq!(lsn2, 2);
-        wal.fsync().unwrap();
 
         assert_eq!(wal.files.len(), 1);
 
@@ -309,8 +420,7 @@ pub mod tests {
         wal.max_file_size = 100;
 
         let big_data = vec![0u8; 150];
-        wal.write_batch(&big_data).unwrap();
-        wal.fsync().unwrap();
+        wal_write(&mut wal, &big_data).unwrap();
 
         assert_eq!(wal.files.len(), 2);
 
@@ -323,12 +433,10 @@ pub mod tests {
         let mut wal = WalWriter::new(dir.path().to_path_buf()).unwrap();
         wal.max_file_size = 10;
 
-        wal.write_batch(b"aaaaaaaaaa_plus").unwrap();
-        wal.fsync().unwrap();
+        wal_write(&mut wal, b"aaaaaaaaaa_plus").unwrap();
         assert_eq!(wal.files.len(), 2);
 
-        wal.write_batch(b"bbbbbbbbbb_plus").unwrap();
-        wal.fsync().unwrap();
+        wal_write(&mut wal, b"bbbbbbbbbb_plus").unwrap();
         assert_eq!(wal.files.len(), 3);
 
         let first_max_lsn = wal.files[0].max_lsn;
@@ -345,8 +453,7 @@ pub mod tests {
         let dir = tempfile::tempdir().unwrap();
         let mut wal = WalWriter::new(dir.path().to_path_buf()).unwrap();
 
-        wal.write_batch(b"data").unwrap();
-        wal.fsync().unwrap();
+        wal_write(&mut wal, b"data").unwrap();
 
         wal.delete_below(u64::MAX);
         assert_eq!(wal.files.len(), 1);
@@ -380,5 +487,37 @@ pub mod tests {
         );
 
         wal.assert_invariants();
+    }
+
+    #[test]
+    fn wal_record_roundtrip() {
+        let record = WalRecord::data(Bytes::from("hello world"));
+        let mut buf = Vec::new();
+        record.encode_to(&mut buf).unwrap();
+
+        let decoded = WalRecord::decode_from(&mut &buf[..]).unwrap();
+        assert_eq!(record, decoded);
+    }
+
+    #[test]
+    fn wal_batch_end_roundtrip() {
+        let record = WalRecord::batch_end();
+        let mut buf = Vec::new();
+        record.encode_to(&mut buf).unwrap();
+
+        let decoded = WalRecord::decode_from(&mut &buf[..]).unwrap();
+        assert_eq!(record, decoded);
+    }
+
+    #[test]
+    fn wal_crc_detects_corruption() {
+        let record = WalRecord::data(Bytes::from("test data"));
+        let mut buf = Vec::new();
+        record.encode_to(&mut buf).unwrap();
+
+        buf[HEADER_SIZE + 2] ^= 0xFF;
+
+        let result = WalRecord::decode_from(&mut &buf[..]);
+        assert!(result.is_err());
     }
 }
